requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==19.3.0
audio-metadata==0.11.1
backports.zoneinfo==0.2.1
bidict==0.22.1
billiard==3.6.4.0
bitstruct==8.17.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-admin-rangefilter==0.9.0
django-cors-headers==3.14.0
django-encrypted-model-fields==0.6.5
django-environ==0.10.0
django-extensions==3.2.1
django-import-export==3.1.0
django-redis==5.2.0
django-rest-framework==0.1.0
django-safedelete==1.3.1
django-searchable-encrypted-fields==0.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
kombu==5.2.4
MarkupPy==1.14
more-itertools==8.14.0
mutagen==1.46.0
odfpy==1.4.1
openpyxl==3.1.2
pendulum==2.1.2
phonenumbers==8.13.7
Pillow==9.4.0
pprintpp==0.4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.17
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tbm-utils==2.6.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
